requirements.txt
Flask[async]==3.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
werkzeug==3.0.1
python-dotenv==1.0.0
prometheus_flask_exporter==0.23.0
python-logstash==0.4.8
Flask-GraphQL==2.0.1
graphene==2.1.9
flask_cors==4.0.0
timeout_decorator==0.5.0
circuitbreaker==2.0.0
flasgger==0.9.7.1
flask_sqlalchemy==3.1.1